Easblock testcase

+++ easybuild/tools/module_generator.py
@@ -167,8 +167,9 @@ if { ![is-loaded %(name)s/%(version)s] } {
         Desconstructor: clean up temporary directory used for fake modules, if any.
         """
         if self.fake:
-            log.info("Cleaning up fake modules dir %s" % self.tmpdir)
             try:
-                shutil.rmtree(self.tmpdir)
+                if os.path.exists(self.tmpdir):
+                    log.info("Cleaning up fake modules dir %s" % self.tmpdir)
+                    shutil.rmtree(self.tmpdir)
             except OSError, err:
                 log.exception("Cleaning up fake module dir failed: %s" % err)
